requirements.txt
pandas
requests
sqlalchemy
pymysql
python-decouple